longarr phase 1: widen heap & MemoryModel byte counts to uint64_t
First PR of the 64-bit-arrays project (plan: swift-rolling-pelican). Lifts
the heap layer's allocation-size arithmetic from uint32_t to uint64_t so
single allocations above 4GB become legal. Array / Table struct widening
and the daslang surface (long_length, arr[i64], int64 reserve/resize) land
in PR #2.

Headers widened (uint32_t -> uint64_t):
  * AnyHeapAllocator virtuals: impl_allocate / impl_free / impl_reallocate
    / mark / isOwnPtr / isValidPtr / setInitialSize / breakOnFree
    / impl_allocateIterator / impl_allocateString / impl_freeString / intern.
    getInitialSize() returns int64_t.
  * All four concrete overrides (Persistent/Linear Heap+String allocators)
    plus ConstStringAllocator.
  * MemoryModel allocate / free / reallocate / setInitialSize
    / isOwnPtr / isAllocatedPtr / bytesAllocated / maxBytesAllocated,
    plus totalAllocated / maxAllocated / initialSize fields and the
    bigStuff (and DAS_SANITIZER deletedBigStuff) value type.
  * LinearChunkAllocator allocate / free / reallocate / setInitialSize
    plus initialSize / unadjustedInitialSize fields.
  * Context allocate / free / reallocate / allocateIterator
    / allocateString / freeString and throw_out_of_memory.

Stays uint32_t by design (per-stack, small-bucket, or chunk-bounded):
  * StackAllocator (per-stack scope, bounded by setup-time stackSize).
  * Shoe / Deck (small-allocation bucket, hard-capped at 256B).
  * HeapChunk / LinearChunkAllocator::grow / MemoryModel::grow
    (chunk-byte and deck-count helpers, never touch the >4GB path).
  * NodePrefix::magic + padd0 (drops padd1 to keep sizeof == sizeof(vec4f)).
  * Internal intern paths cast inbound uint64_t length to uint32_t for
    StrHashEntry (interned strings stay <4GB by construction).

Other:
  * DAS_PAGE_GC_MASK moved from bit 31 to bit 63 so size-and-mark packing
    in bigStuff survives the widened value type.
  * 3 new C-API entries -- das_context_allocate_i64 / _reallocate_i64
    / _free_i64. Legacy uint32_t entries unchanged (ABI-stable, forward
    via zero-extending implicit upcast).
  * tests-cpp/small/test_heap_64bit.cpp -- 3 TEST_CASEs / 286 assertions
    covering _i64 round-trip, legacy uint32 forward path, and a 5GB
    allocation gated on DASLANG_HUGE_HEAP_TESTS=1.
  * tests/long_array_table/_milestone.das (leading underscore = dastest
    discovery skip per feedback_dastest_underscore_skip) -- compile-gap
    scaffold; today fails to compile on master, exercises the full
    long_*/arr[i64]/long_iter_range surface that PR #2 introduces.

Verification: libDaScript builds clean; cascade callers in src/builtin
and tests-cpp resolve via implicit uint32->uint64 upcast at the call
site. 8819/8819 .das tests pass. tests-cpp-small new TEST_CASEs all
green. builtin_table production bench canaries within 10% gate vs
Phase 0 baseline (most show -15% to -28% improvements on the cleaner
post-merge machine state; one +6.3% on the 8 ns/op overwrite_same_key
canary, at the 1.1% CV noise floor).

+++ include/daScript/simulate/simulate.h
@@ -356,7 +356,7 @@ namespace das
         void onReallocate ( void * ptr, uint64_t size, void * newPtr, uint64_t newSize, const LineInfo & at );
         void onFree ( void * ptr, const LineInfo & at );
 
-        __forceinline char * allocateIterator ( uint32_t size, const char * iterName, const LineInfo * at ) {
+        __forceinline char * allocateIterator ( uint64_t size, const char * iterName, const LineInfo * at ) {
             auto aptr = heap->impl_allocateIterator(size, iterName);
             if ( !aptr ) throw_out_of_memory(false, size + 16, at);
             if ( instrumentAllocations ) onAllocate(aptr - 16, size + 16, at ? *at : LineInfo());
@@ -368,26 +368,26 @@ namespace das
             heap->impl_freeIterator(ptr);
         }
 
-        __forceinline char * allocate ( uint32_t size, const LineInfo * at = nullptr ) {
+        __forceinline char * allocate ( uint64_t size, const LineInfo * at = nullptr ) {
             auto aptr = heap->impl_allocate(size);
             if ( !aptr && size ) throw_out_of_memory(false, size, at);
             if ( instrumentAllocations ) onAllocate(aptr, size, at ? *at : LineInfo());
             return aptr;
         }
 
-        __forceinline char * reallocate ( char * ptr, uint32_t oldSize, uint32_t size, const LineInfo * at ) {
+        __forceinline char * reallocate ( char * ptr, uint64_t oldSize, uint64_t size, const LineInfo * at ) {
             auto aptr = heap->impl_reallocate(ptr, oldSize, size);
             if ( !aptr && size ) throw_out_of_memory(false, size, at);
             if ( instrumentAllocations ) onReallocate(ptr, oldSize, aptr, size, at ? *at : LineInfo());
             return aptr;
         }
 
-        __forceinline void free ( char * ptr, uint32_t size, const LineInfo * at = nullptr ) {
+        __forceinline void free ( char * ptr, uint64_t size, const LineInfo * at = nullptr ) {
             if ( instrumentAllocations ) onFree(ptr, at ? *at : LineInfo());
             heap->impl_free(ptr, size);
         }
 
-        __forceinline char * allocateString ( const char * text, uint32_t length, const LineInfo * at, bool tempString = false ) {
+        __forceinline char * allocateString ( const char * text, uint64_t length, const LineInfo * at, bool tempString = false ) {
             auto astr = stringHeap->impl_allocateString(this, text, length, at);
             if ( !astr && length ) throw_out_of_memory(true, length+1, at);
             if ( instrumentAllocations ) onAllocateString(astr, length, tempString, at ? *at : LineInfo());
@@ -395,18 +395,18 @@ namespace das
         }
 
         __forceinline char * allocateString ( const string & str, const LineInfo * at, bool tempString = false ) {
-            auto astr = stringHeap->impl_allocateString(this, str.c_str(), uint32_t(str.size()), at);
-            if ( !astr && str.size() ) throw_out_of_memory(true, uint32_t(str.size()+1), at);
+            auto astr = stringHeap->impl_allocateString(this, str.c_str(), uint64_t(str.size()), at);
+            if ( !astr && str.size() ) throw_out_of_memory(true, uint64_t(str.size()+1), at);
             if ( instrumentAllocations ) onAllocateString(astr, str.size(), tempString, at ? *at : LineInfo());
             return astr;
         }
 
-        __forceinline char * allocateTempString ( const char * text, uint32_t length, const LineInfo * at ) {
+        __forceinline char * allocateTempString ( const char * text, uint64_t length, const LineInfo * at ) {
             return allocateString(text, length, at, /*temp*/true);
         }
 
-        __forceinline bool freeString ( char * ptr, uint32_t length, const LineInfo * at, bool tempString = false ) {
-            uint32_t size = length + 1;
+        __forceinline bool freeString ( char * ptr, uint64_t length, const LineInfo * at, bool tempString = false ) {
+            uint64_t size = length + 1;
             size = (size + 15) & ~15;
             if (stringHeap->isOwnPtr(ptr, size)) {
                 if ( instrumentAllocations ) onFreeString(ptr, tempString, at ? *at : LineInfo());
@@ -418,7 +418,7 @@ namespace das
 
         __forceinline void freeTempString ( char * ptr, const LineInfo * at ) {
             if ( stringHeap->isIntern() ) return;
-            if ( stringDisposeQue ) freeString(stringDisposeQue,(uint32_t)strlen(stringDisposeQue),at, /*temp*/true);
+            if ( stringDisposeQue ) freeString(stringDisposeQue,(uint64_t)strlen(stringDisposeQue),at, /*temp*/true);
             stringDisposeQue = ptr;
         }
 
@@ -502,7 +502,7 @@ namespace das
         DAS_NORETURN_PREFIX void throw_error_at ( const LineInfo * at, DAS_FORMAT_STRING_PREFIX const char * message, ... ) DAS_NORETURN_SUFFIX DAS_FORMAT_PRINT_ATTRIBUTE(3,4);
         DAS_NORETURN_PREFIX void throw_fatal_error ( const char * message, const LineInfo & at ) DAS_NORETURN_SUFFIX;
         DAS_NORETURN_PREFIX void rethrow () DAS_NORETURN_SUFFIX;
-        DAS_NORETURN_PREFIX void throw_out_of_memory ( bool stringHeap, uint32_t size, const LineInfo * at=nullptr ) DAS_NORETURN_SUFFIX;
+        DAS_NORETURN_PREFIX void throw_out_of_memory ( bool stringHeap, uint64_t size, const LineInfo * at=nullptr ) DAS_NORETURN_SUFFIX;
 
         __forceinline SimFunction * getFunction ( int index ) const {
             return (index>=0 && index<totalFunctions) ? functions + index : nullptr;
